add: final style, responsivity and all API requests

+++ src/components/TableBody/TableBody.tsx
@@ -1,22 +1,29 @@
-
-import { Pessoa } from '@/types/Pessoa'
+'use client'
+import { Person } from '@/types/Person'
 import styles from './tBody.module.css';
 import { MdDelete } from 'react-icons/md';
 import { AiFillEdit } from 'react-icons/ai';
 import api from '@/services/api';
 import Link from 'next/link';
+import React from 'react';
 
 type Props = {
-  data: Pessoa[] | undefined;
+  data: Person[] | undefined;
 }
 
 const TableBody = ({ data }: Props) => {
-  const handleDeletePeople = async (id: number | undefined, nome: string) => {
-    try {
-      await api.delete(`/pessoa/${id}`);
-      return window.alert(`${nome} foi deletado com sucesso>!`);
-    } catch (error) {
-      return window.alert(error);
+  const handleDeletePerson = async (id: number | undefined, nome: string) => {
+
+    if (confirm(`Deseja mesmo excluir o registro de ${nome}? `) == true) {
+      try {
+        await api.delete(`/pessoa/${id}`);
+        return window.alert(`${nome} foi deletado com sucesso!`);
+      } catch (error) {
+        return window.alert(error);
+      }
+    }
+    else {
+      return window.alert('O registro não foi excluído!');
     }
   }
   return (
@@ -24,12 +31,15 @@ const TableBody = ({ data }: Props) => {
       {
         data?.map((pessoa) => {
           return (
-            <tr className={styles.tBody} key={pessoa.id}>
+            <tr key={pessoa.id}>
               <td>{pessoa.tipoPessoa}</td>
               <td>{pessoa.nome_RazaoSocial}</td>
               <td>{pessoa.cpF_CNPJ}</td>
               <td>{pessoa.telefone}</td>
               <td>{pessoa.email}</td>
+              <td>
+                {pessoa.endereco.map((endereco) => <td className={styles.td}>{endereco.tipoEndereco}</td>)}
+              </td>
               <td>
                 {pessoa.endereco.map((endereco) => <td className={styles.td}>{endereco.endereco}</td>)}
               </td>
@@ -58,7 +68,7 @@ const TableBody = ({ data }: Props) => {
               </td>
               <td>
                 <MdDelete className={styles.icon}
-                  onClick={() => handleDeletePeople(pessoa.id, pessoa.nome_RazaoSocial)}
+                  onClick={() => handleDeletePerson(pessoa.id, pessoa.nome_RazaoSocial)}
                   size={32}
                   color='red' />
               </td>
